backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
cachetools==5.5.2
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
Django==5.0.14
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-timezone-field==7.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring_parser==0.17.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-genai==1.14.0
greenlet==3.3.2
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
oauthlib==3.3.1
openai==2.24.0
packaging==26.0
pillow==12.1.1
playwright==1.58.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
pytest==9.0.2
pytest-base-url==2.1.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
redis==5.0.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stripe==14.4.0
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
whitenoise==6.6.0
